Android UI: Fix some linter warnings
This commit fixes some Java linter warnings.
It also makes the MobileRT work on Android 11 (or Android API 30), by
ignoring some errors from `errno` that might occur in C++ native code.

+++ app/src/main/java/puscas/mobilertapp/MainActivity.java
@@ -661,8 +661,8 @@ public final class MainActivity extends Activity {
         vto.addOnGlobalLayoutListener(() -> {
             log.info("initializePickerResolutions");
             resetErrno();
-            final double widthView = (double) this.drawView.getWidth();
-            final double heightView = (double) this.drawView.getHeight();
+            final double widthView = this.drawView.getWidth();
+            final double heightView = this.drawView.getHeight();
 
             final String[] resolutions = IntStreams.rangeClosed(2, maxSizes)
                 // double.class::cast method throws `ClassCastException: Integer cannot be cast to double`
